+++ properties/Directory/MapFiles.php
@@ -33,7 +33,7 @@ final class MapFiles implements Property
 
     public function applicableTo(object $directory): bool
     {
-        return !$directory->files()->empty();
+        return !$directory->all()->empty();
     }
 
     public function ensureHeldBy(Assert $assert, object $directory): object
@@ -46,14 +46,14 @@ final class MapFiles implements Property
             ->same($directory2);
         $assert->same($directory->name()->toString(), $directory2->name()->toString());
         $assert
-            ->expected($directory->files())
+            ->expected($directory->all())
             ->not()
-            ->same($directory2->files());
-        $assert->same($directory->files()->size(), $directory2->files()->size());
+            ->same($directory2->all());
+        $assert->same($directory->all()->size(), $directory2->all()->size());
         $assert->same(
             [$this->file->content()],
             $directory2
-                ->files()
+                ->all()
                 ->map(static fn($file) => $file->content())
                 ->distinct()
                 ->toList(),